--- a/src/grammar.json
+++ b/src/grammar.json
@@ -4517,71 +4517,6 @@
         ]
       }
     },
-    "_legacy_pattern": {
-      "type": "CHOICE",
-      "members": [
-        {
-          "type": "SEQ",
-          "members": [
-            {
-              "type": "SYMBOL",
-              "name": "_legacy_pattern"
-            },
-            {
-              "type": "SYMBOL",
-              "name": "logical_or_operator"
-            },
-            {
-              "type": "SYMBOL",
-              "name": "_legacy_pattern"
-            }
-          ]
-        },
-        {
-          "type": "SEQ",
-          "members": [
-            {
-              "type": "SYMBOL",
-              "name": "_legacy_pattern"
-            },
-            {
-              "type": "SYMBOL",
-              "name": "logical_and_operator"
-            },
-            {
-              "type": "SYMBOL",
-              "name": "_legacy_pattern"
-            }
-          ]
-        },
-        {
-          "type": "SEQ",
-          "members": [
-            {
-              "type": "CHOICE",
-              "members": [
-                {
-                  "type": "SYMBOL",
-                  "name": "relational_operator"
-                },
-                {
-                  "type": "SYMBOL",
-                  "name": "equality_operator"
-                }
-              ]
-            },
-            {
-              "type": "SYMBOL",
-              "name": "_real_expression"
-            }
-          ]
-        },
-        {
-          "type": "SYMBOL",
-          "name": "_unary_pattern"
-        }
-      ]
-    },
     "_unary_pattern": {
       "type": "CHOICE",
       "members": [
@@ -4613,10 +4548,6 @@
         {
           "type": "SYMBOL",
           "name": "variable_pattern"
-        },
-        {
-          "type": "SYMBOL",
-          "name": "_parenthesized_pattern"
         },
         {
           "type": "SYMBOL",
@@ -4908,7 +4839,7 @@
         },
         {
           "type": "SYMBOL",
-          "name": "_legacy_pattern"
+          "name": "_pattern"
         },
         {
           "type": "STRING",
@@ -12386,9 +12317,6 @@
     ],
     [
       "_pattern"
-    ],
-    [
-      "_legacy_pattern"
     ],
     [
       "constructor_tearoff",