This PR is just a workaround fix for #46.
all test passed

## Performance
### Test source
```sh
$ cat queries/good.scm
(if_statement)
$ cat queries/bad.scm
(if_statement (block))
$ cat empty.dart # empty file
```

### before changes
```sh
$ time tree-sitter query queries/good.scm empty.dart
tree-sitter query queries/good.scm empty.dart  0.01s user 0.00s system 61% cpu 0.021 total

$ time tree-sitter query queries/bad.scm empty.dart
tree-sitter query queries/bad.scm empty.dart  1.23s user 0.01s system 99% cpu 1.245 total
```

### after changes

```sh
$ time tree-sitter query queries/good.scm empty.dart
tree-sitter query queries/good.scm empty.dart  0.01s user 0.00s system 74% cpu 0.017 total

$ time tree-sitter query queries/bad.scm empty.dart
tree-sitter query queries/bad.scm empty.dart  0.14s user 0.00s system 97% cpu 0.148 total
```