@TimWhiting 2023-06-13
Thanks, I ended up commenting it out rather than removing it completely, we need to figure out a way to reinsert it eventually since it allows you to control precedence of patterns. As there is very little such code out there currently, I'll ignore it for now. It would be great though if we could get a failing test added so that we can work on eventually reinserting this piece of the grammar.

Anyways, thanks for the PR. I'll add your name to the contributors in package.json, even though I made the changes myself.