@TimWhiting 2023-06-13
@theHamsta

This change is incorporated, and you can update nvim to the latest commit.